requirements.txt
pygame>=2.5.0
PyYAML>=6.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
